Widgets returns indicators 2/2
