+++ lib/controllers/widget_controller.js
@@ -50,6 +50,30 @@ var WidgetController = {
     Widget.findOne({
         _id: id
     }, fn);
+  },
+
+  getPublicJson: function(widget, fn){
+    var publicJson = widget.getPublicJson();
+    if (widget.indicator) {
+      IndicatorController.findById(widget.indicator, function(indicator) {
+        IndicatorController.getValues(widget.indicator, function(values){
+          publicJson.indicator = {
+            source: {
+              provider: {
+                name: indicator.provider,
+                token: ""
+              },
+              id: indicator.source
+            },
+            values: values
+          };
+
+          fn(null, publicJson);  
+        });
+      });
+    } else {
+      fn(null, publicJson);
+    }
   }
 };
 module.exports = WidgetController;
