+++ lib/routes.js
@@ -3,6 +3,7 @@ module.exports = Routes;
 /************
  * Requires *
  ************/
+var async = require('async');
 
 //Controllers
 UserController = require('./controllers/user_controller');
@@ -253,10 +254,9 @@ function widgets(req, res, next) {
         message: "You dont have any widget."
       });
     }
-
-    res.json(widgets.map(function(widget) {
-      return widget.getPublicJson();
-    }));
+    async.map(widgets, WidgetController.getPublicJson, function (e, result) {
+      res.json(result);
+    });
   });
 }
 
@@ -286,7 +286,9 @@ function createWidget(req, res, next) {
     if (err) {
       return res.status(500).json(err);
     }
-    res.status(201).json(widget.getPublicJson());
+    WidgetController.getPublicJson(widget, function(err, publicJson){
+      res.status(201).json(publicJson);
+    });
   });
 }
 
@@ -303,26 +305,9 @@ function widget(req, res, next) {
         message: "Couldn´t find any widget with this id."
       });
     }
-    var publicJson = widget.getPublicJson();
-    if (widget.indicator) {
-      IndicatorController.findById(this.indicator, function(indicator) {
-        IndicatorController.getValues(this.indicator, function(values){
-          publicJson.indicator = {
-            source: {
-              provider: {
-                name: indicator.provider,
-                token: ""
-              },
-              id: indicator.source
-            },
-            values: values
-          }
-          res.json(publicJson);  
-        });
-      });
-    } else {
+    WidgetController.getPublicJson(widget, function(err, publicJson){
       res.json(publicJson);
-    }
+    });
   });
 }
