+++ package.json
@@ -22,6 +22,7 @@
     "expect.js": "^0.3.1"
   },
   "dependencies": {
+    "async": "^1.4.0",
     "bcryptjs": "^2.1.0",
     "body-parser": "^1.13.2",
     "client-sessions": "^0.7.0",
